# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/llm/process": {
            "post": {
                "tags": [
                    "llm"
                ],
                "summary": "Process With Llm",
                "description": "Unified LLM processing endpoint that can handle various prompts and input sources",
                "operationId": "process_with_llm_llm_process_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LLMRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/llm/prompts": {
            "get": {
                "tags": [
                    "llm"
                ],
                "summary": "Get Prompts",
                "description": "Get list of available named prompts",
                "operationId": "get_prompts_llm_prompts_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptsResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/llm/list": {
            "get": {
                "tags": [
                    "llm"
                ],
                "summary": "List Results",
                "description": "List all saved LLM results",
                "operationId": "list_results_llm_list_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMListResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/llm/file/{result_id}": {
            "get": {
                "tags": [
                    "llm"
                ],
                "summary": "Get Result File",
                "description": "Get LLM result file content by ID",
                "operationId": "get_result_file_llm_file__result_id__get",
                "parameters": [
                    {
                        "name": "result_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Result Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "llm"
                ],
                "summary": "Delete Result File",
                "description": "Delete LLM result file by ID",
                "operationId": "delete_result_file_llm_file__result_id__delete",
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "result_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Result Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/llm/extract": {
            "post": {
                "tags": [
                    "llm"
                ],
                "summary": "Extract Learning",
                "description": "Convenience endpoint for learning extraction (equivalent to learning/scrape)",
                "operationId": "extract_learning_llm_extract_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExtractRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/llm/analyze": {
            "post": {
                "tags": [
                    "llm"
                ],
                "summary": "Analyze Content",
                "description": "Convenience endpoint for content analysis (equivalent to analysis/analyze)",
                "operationId": "analyze_content_llm_analyze_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/app__llm__models__AnalyzeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
        "/llm/custom": {
            "post": {
                "tags": [
                    "llm"
                ],
                "summary": "Custom Processing",
                "description": "Custom prompt processing endpoint",
                "operationId": "custom_processing_llm_custom_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LLMResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "APIKeyHeader": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomRequest": {
                "properties": {
                    "custom_prompt": {
                        "type": "string",
                        "title": "Custom Prompt"
                    },
                    "query_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Url"
                    },
                    "query_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Text"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "default": "gpt-4o"
                    },
                    "save_result": {
                        "type": "boolean",
                        "title": "Save Result",
                        "default": false
                    },
                    "output_format": {
                        "$ref": "#/components/schemas/OutputFormat",
                        "default": "text"
                    }
                },
                "type": "object",
                "required": [
                    "custom_prompt"
                ],
                "title": "CustomRequest",
                "description": "Request model for custom processing endpoint"
            },
            "ExtractRequest": {
                "properties": {
                    "query_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Url"
                    },
                    "query_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Text"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "default": "gpt-4o"
                    },
                    "save_result": {
                        "type": "boolean",
                        "title": "Save Result",
                        "default": true
                    }
                },
                "type": "object",
                "title": "ExtractRequest",
                "description": "Request model for extract convenience endpoint"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "LLMListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/LLMListItem"
                        },
                        "type": "array",
                        "title": "Data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "LLMListResponse",
                "description": "Response model for LLM results list endpoint"
            },
            "LLMRequest": {
                "properties": {
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "description": "LLM model to use",
                        "default": "gpt-4o"
                    },
                    "prompt": {
                        "type": "string",
                        "title": "Prompt",
                        "description": "Named prompt (e.g., 'extract_learning', 'analyze_summary') or custom prompt text"
                    },
                    "query_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Text",
                        "description": "Direct text input to process"
                    },
                    "query_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Url",
                        "description": "URL to scrape and process"
                    },
                    "save_result": {
                        "type": "boolean",
                        "title": "Save Result",
                        "description": "Whether to save result to file and return URL",
                        "default": false
                    },
                    "output_format": {
                        "$ref": "#/components/schemas/OutputFormat",
                        "description": "Output format for the response",
                        "default": "text"
                    },
                    "custom_title": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Custom Title",
                        "description": "Custom title for saved results"
                    },
                    "metadata": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Metadata",
                        "description": "Additional metadata for saved results"
                    }
                },
                "type": "object",
                "required": [
                    "prompt"
                ],
                "title": "LLMRequest",
                "description": "Request model for unified LLM endpoint",
                "example": {
                    "model": "gpt-4o",
                    "output_format": "markdown",
                    "prompt": "analyze_summary",
                    "query_url": "https://example.com/article",
                    "save_result": true
                }
            },
            "LLMResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Response status (success/error)"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "description": "Response message"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data",
                        "description": "Result data - either LLMResult dict, URL string, or direct content"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message"
                ],
                "title": "LLMResponse",
                "description": "Standard response model for LLM operations"
            },
...
            "OutputFormat": {
                "type": "string",
                "enum": [
                    "markdown",
                    "json",
                    "text"
                ],
                "title": "OutputFormat",
                "description": "Output format options for LLM responses"
            },
...
            "PromptInfo": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Prompt identifier"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "description": "Description of what the prompt does"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category",
                        "description": "Category of the prompt (analysis, learning, etc.)"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "category"
                ],
                "title": "PromptInfo",
                "description": "Information about available prompts"
            },
            "PromptsResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "data": {
                        "items": {
                            "$ref": "#/components/schemas/PromptInfo"
                        },
                        "type": "array",
                        "title": "Data"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "title": "PromptsResponse",
                "description": "Response model for available prompts endpoint"
            },
...
            "app__llm__models__AnalyzeRequest": {
                "properties": {
                    "analysis_type": {
                        "type": "string",
                        "title": "Analysis Type",
                        "default": "summary"
                    },
                    "query_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Url"
                    },
                    "query_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Query Text"
                    },
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "default": "gpt-4o"
                    },
                    "save_result": {
                        "type": "boolean",
                        "title": "Save Result",
                        "default": true
                    }
                },
                "type": "object",
                "title": "AnalyzeRequest",
                "description": "Request model for analyze convenience endpoint"
            }
        },
        "securitySchemes": {
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key"
            }
        }
    }
}